+++ src/tools/unmunch.cxx
@@ -202,7 +202,7 @@ int parse_aff_file(FILE* afflst) {
             case 3: {
               numents = atoi(piece);
               if ((numents < 0) ||
-                  ((SIZE_MAX / sizeof(struct affent)) < numents)) {
+                  ((SIZE_MAX / sizeof(struct affent)) < static_cast<size_t>(numents))) {
                 fprintf(stderr, "Error: too many entries: %d\n", numents);
                 numents = 0;
               } else {
